+ SignUp.jsx and SignIn.jsx now uses the same theme folder since they both had indentical themes. - Removed old SignUp and SignIn .js files (I would have moved them to the deprecated folder but it doesn't exist anymore)
