+++ frontend/src/components/authentication/SignUp.jsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 import { Link } from 'react-router-native';
 import TextInput from "../TextInput";
 import { useFormik } from 'formik';
-import theme from "../../themes/signUpTheme";
+import theme from "../../themes/signInAndSingUpTheme.";
 import mainTheme from "../../themes/mainTheme";
 import * as yup from "yup";
