--- frontend/src/themes/SignInTheme.js
@@ -1,37 +1,0 @@
-const theme = {
-      pinkContainer: {
-        marginHorizontal: 18,
-        backgroundColor: "rgba(237,220,249,1)",
-        borderColor: "rgba(187,144,214,1)",
-        borderRadius: 48,
-        borderWidth: 7,
-        alignContent: "center",
-        padding: 10,
-      },
-      arrowContainer: {
-        height: 30,
-        width: 30,
-        margin: 10,
-      },
-      arrow: {
-        fontSize: 30,
-        color: "rgba(75,113,123,1)",
-      },
-
-      errorText: {
-        color: "#d73a4a",
-        marginLeft: 10
-      },
-      inputField: {
-        paddingVertical: 10,
-        backgroundColor: "rgba(255,255,255,1)",
-        borderColor: "rgba(100,152,166,1)",
-        borderRadius: 10,
-        borderWidth: 2,
-        color: "rgba(100,152,166,1)",
-        fontSize: 25,
-      },
-
-};
-
-export default theme;
